requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
PyMySQL==1.1.0
SQLAlchemy==2.0.27
typing-extensions>=4.12
Werkzeug==3.0.1
WTForms==3.1.2
Flask-Bcrypt==1.0.1
cryptography==41.0.7
PyJWT==2.12.1